Exceptions handling discussion
I think the approach in this commit is only partial
Same for most the addSyncTaskListVer8() parsed options + all the catch(Exception e){} empty entries

exp+proof of concept: we edit settings file and change entries for dst_offset and time_diff_allowable_time to 600. Open SMBSync2, import the hacked task list. Do not enter settings, exit SMBSync2. Open the auto-saved task list created on exit: it will contain the 600 value for both the hacked fields.

If we start the sync: setSyncOptionDifferentFileAllowableTime(600) is triggered and getSyncOptionDifferentFileAllowableTime() will always return 600 in the active session unless we go to task settings. The parser Exception won't trigger since 600 is an integer. The sync task will effectively apply 600 seconds time offset. Same for the dst offset. If we edit the task in advanced settings, problem is fixed

Proper fix: whenever you read a settings file, the rule is to account for corruption and user modifications exceptions.
exp (code not exact):
                int set = Integer.parseInt(parm[34];
                // add code to handle exception if it is an invalid string like "zzz"
                if (set != 1 or or 3 or 10) // valid range or values
                    set = 3; // default value
                stli.setSyncOptionDifferentFileAllowableTime(set);

I think all the entries in addSyncTaskListVer8() should be checked at entry, and proper default values set if invalid entries detected.
This is a rule when a program accepts values from a user input or external file

Hope you can fix it as you are used to the code. It would take me weeks to look through all your code to get the proper defaults :-)

Best regards