Implement simpler and more generic onLongTouch handler
This replaces Messages and History move context buttons long touch code by a generic class.

The new onLongTouch Repeater class is much more simpler to implement in code with just the onTouchListener. It can be used on every view that needs a long touch repeated action

The code root is the same as previously, using a handler to the repeatable runner like recommended for Android

Next commits will tweak some aspects